+++ app/db/write/src/commonMain/kotlin/com/moneymanager/database/repository/write/CurrencyWriteRepositoryImpl.kt
@@ -39,7 +39,7 @@ class CurrencyWriteRepositoryImpl(
                         // Allocate an id from the shared `asset` id space, then insert the currency with it.
                         assetWriteQueries.insert()
                         val newId = assetWriteQueries.lastInsertedId().executeAsOne()
-                        writeQueries.insert(newId, code, name, scaleFactor.toLong())
+                        writeQueries.insert(newId, code, name, scaleFactor)
                         // Only a freshly inserted currency records a source (the existing branch keeps its own).
                         database.recordSource(deviceId, EntityType.CURRENCY, newId, 1L, source)
                         CurrencyId(newId)
